chore: refactor out processAssetParams to utils
