+++ ui/pages/asset/asset.tsx
@@ -1,5 +1,5 @@
 import { Nft } from '@metamask/assets-controllers';
-import { CaipChainId, Hex, isCaipChainId } from '@metamask/utils';
+import { CaipChainId, Hex } from '@metamask/utils';
 import React, { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { Navigate, useParams, useLocation } from 'react-router-dom';
@@ -11,6 +11,7 @@ import { DEFAULT_ROUTE } from '../../helpers/constants/routes';
 import { getTokenByAccountAndAddressAndChainId } from '../../selectors/assets';
 import NativeAsset from './components/native-asset';
 import TokenAsset from './components/token-asset';
+import { processAssetParams } from './util';
 
 type LocationState = {
   token?: {
@@ -22,35 +23,6 @@ type LocationState = {
     isNative?: boolean;
     decimals: number;
   };
-};
-
-/**
- * Firefox and Chrome process the asset params differently due to how they handle decoding fragments.
- * E.g. With a route of `/asset/solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/solana%3A5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp%2Ftoken%3AXXX`
- * (where the solana%3A5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp%2Ftoken%3AXXX is an encoded version of the asset id)
- *
- * - Chrome will decode the above path as `{chainId}/{asset}`
- * - Chrome will decode the `asset` param as solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/token:XXX
- * - Chrome will therefore leave the `id` param as undefined.
- *
- * - Firefox will decode the above path as `{chainId}/{asset}/{id}`
- * - Firefox will decode the `asset` param as solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp
- * - Firefox will therefore leave the `id` param as token:XXX
- *
- * @param params - route params
- * @param params.chainId
- * @param params.asset
- * @param params.id
- * @returns
- */
-export const processAssetParams = (
-  params: Partial<{ chainId: Hex | CaipChainId; asset: string; id: string }>,
-) => {
-  const { chainId, asset, id } = params;
-  const isCaipChain = chainId ? isCaipChainId(chainId) : false;
-  const rawAsset = isCaipChain && asset && id ? `${asset}/${id}` : asset;
-  const decodedAsset = rawAsset ? decodeURIComponent(rawAsset) : undefined;
-  return { chainId, asset, id, decodedAsset };
 };
 
 const Asset = () => {
