+++ ui/pages/asset/util.ts
@@ -1,7 +1,36 @@
 import { SUPPORTED_CHAIN_IDS } from '@metamask/assets-controllers';
-import { Hex, assert } from '@metamask/utils';
+import { CaipChainId, Hex, assert, isCaipChainId } from '@metamask/utils';
 import { Duration } from 'luxon';
 import { PriceApiTimePeriod } from './types/PriceApiTimePeriod';
+
+/**
+ * Firefox and Chrome process the asset params differently due to how they handle decoding fragments.
+ * E.g. With a route of `/asset/solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/solana%3A5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp%2Ftoken%3AXXX`
+ * (where the solana%3A5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp%2Ftoken%3AXXX is an encoded version of the asset id)
+ *
+ * - Chrome will decode the above path as `{chainId}/{asset}`
+ * - Chrome will decode the `asset` param as solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp/token:XXX
+ * - Chrome will therefore leave the `id` param as undefined.
+ *
+ * - Firefox will decode the above path as `{chainId}/{asset}/{id}`
+ * - Firefox will decode the `asset` param as solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp
+ * - Firefox will therefore leave the `id` param as token:XXX
+ *
+ * @param params - route params
+ * @param params.chainId
+ * @param params.asset
+ * @param params.id
+ * @returns
+ */
+export const processAssetParams = (
+  params: Partial<{ chainId: Hex | CaipChainId; asset: string; id: string }>,
+) => {
+  const { chainId, asset, id } = params;
+  const isCaipChain = chainId ? isCaipChainId(chainId) : false;
+  const rawAsset = isCaipChain && asset && id ? `${asset}/${id}` : asset;
+  const decodedAsset = rawAsset ? decodeURIComponent(rawAsset) : undefined;
+  return { chainId, asset, id, decodedAsset };
+};
 
 /** Formats a datetime in a short human readable format like 'Feb 8, 12:11 PM' */
 export const getShortDateFormatter = () =>
